feat(auto-mode): implement continuous trajectory with CW/CCW direction control
feat: resolve Auto-Mode trajectory logic and continuous angle remapping

* Base system direction (CW/CCW) successfully integrated.
* Next step: Fix UI feedback scaling & Re-tune PID for Degree units.

+++ Core/Inc/datatype.h
@@ -177,7 +177,7 @@ typedef struct {
 	Direction_t Direction[9] ;
 	uint8_t Number_of_target ;
 	GripperENA_t Gripper_Ena;
-	int8_t Gripper_sequence[10] ; // Gripper Sequence
+	int16_t Gripper_sequence[10] ; // Gripper Sequence
 } Pick_Place_t ;
 // P2P
 typedef struct {
